+++ functions/index.js
@@ -21623,3 +21623,4 @@ exports.stripeWebhook = functions
 
   res.json({ received: true });
 });
+